fix windows compilation

+++ src/listener.zig
@@ -33,15 +33,16 @@ pub fn listen(comptime S: type, httpz_allocator: Allocator, app_allocator: Alloc
 
 	var no_delay = true;
 	const address = blk: {
-		if (config.unix_path) |unix_path| {
-			no_delay = false;
-			std.fs.deleteFileAbsolute(unix_path) catch {};
-			break :blk try net.Address.initUnix(unix_path);
-		} else {
-			const listen_port = config.port.?;
-			const listen_address = config.address.?;
-			break :blk try net.Address.parseIp(listen_address, listen_port);
+		if (comptime builtin.os.tag != .windows) {
+			if (config.unix_path) |unix_path| {
+				no_delay = false;
+				std.fs.deleteFileAbsolute(unix_path) catch {};
+				break :blk try net.Address.initUnix(unix_path);
+			}
 		}
+		const listen_port = config.port.?;
+		const listen_address = config.address.?;
+		break :blk try net.Address.parseIp(listen_address, listen_port);
 	};
 	try socket.listen(address);
